把那个箭头搞定了，再封装一下就行了。 时隔太久我自己都忘了干了啥了。

+++ Assets/Scripts/Hotfix/Tile/EnemyTile.cs
@@ -11,7 +11,7 @@ namespace Hotfix
     /// <summary>
     /// 进入此格时，进行战斗
     /// </summary>
-    [Tile((int)TileTypeEnum.Enemy)]
+    [Tile(TileTypeEnum.Enemy)]
     public class EnemyTile : Tile
     {
         public override void OnClick()
